ReName: lib/classes/ -> classes/
