+++ classes/User.tsx
@@ -1,10 +1,10 @@
-import {userStatus} from '@/lib/types';
+import {userStatusType} from '@/types/user';
 
 export class User {
   id: string; // UUID
   name: string; // max 10 characters
   profileImage: string;
-  currentStatus: userStatus;
+  currentStatus: userStatusType;
   introduction: string; // max 50 characters
   friendList: string[]; // userId[]
   blockList: string[]; // userId[]
